refactor: restructure Blockaid keysign payload scanning by separating logic into dedicated functions and files
