+++ core/ui/vault/send/verify/SendVerify.tsx
@@ -33,7 +33,7 @@ import { FC } from 'react'
 import { useTranslation } from 'react-i18next'
 import styled from 'styled-components'
 
-import { BlockaidKeysignPayloadScan } from '../../../chain/security/blockaid/keysignPayload'
+import { BlockaidKeysignPayloadScan } from '../../../chain/security/blockaid/keysignPayload/scan'
 import { BlockaidTxStatusContainer } from '../../../chain/security/blockaid/tx/statusContainer'
 import { useIsBlockaidEnabled } from '../../../storage/blockaid'
 import { useSendTxKeysignPayloadQuery } from '../state/useSendTxKeysignPayloadQuery'
